moved SearchMoviesResults to index.js
